set initial value
